+++ CHANGELOG.md
@@ -11,11 +11,11 @@ All notable changes to this project will be documented in this file. Take a look
 #### Shared
 
 * Implementation of the [W3C Accessibility Metadata Display Guide](https://w3c.github.io/publ-a11y/a11y-meta-display-guide/2.0/guidelines/) specification to facilitate displaying accessibility metadata to users. [See the dedicated user guide](docs/Guides/Accessibility.md).
+* Support for starting from a progression in the HTML content iterator.
 
 #### Navigator
 
 * A new `InputObserving` API has been added to enable more flexible gesture recognition and support for mouse pointers. [See the dedicated user guide](docs/Guides/Navigator/Input.md).
-* Support for starting from a progression in the HTML content iterator.
 
 ### Fixed