+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 All notable changes to this project will be documented in this file. Take a look at [the migration guide](docs/Migration%20Guide.md) to upgrade between two major versions.
 
-**Warning:** Features marked as *alpha* may change or be removed in a future release without notice. Use with caution.
+<!-- ## [Unreleased] -->
 
-## [Unreleased]
+## [3.3.0]
 
 ### Added
 
@@ -930,3 +930,4 @@ progression. Now if no reading progression is set, the `effectiveReadingProgress
 [3.0.0-beta.2]: https://github.com/readium/swift-toolkit/compare/3.0.0-beta.1...3.0.0-beta.2
 [3.1.0]: https://github.com/readium/swift-toolkit/compare/3.0.0...3.1.0
 [3.2.0]: https://github.com/readium/swift-toolkit/compare/3.1.0...3.2.0
+[3.3.0]: https://github.com/readium/swift-toolkit/compare/3.2.0...3.3.0